fix num_periods not cast to int

+++ backend/pipeline/predict.py
@@ -6,7 +6,7 @@ from fbprophet import Prophet
 
 def predict(data):
     '''Return a dict like {variable_name: past_and_future_dataframe}'''
-    num_periods = np.clip(len(data) * 0.25, 4, 256)
+    num_periods = int(np.clip(len(data) * 0.25, 4, 256))
     result = {}
     for variable in data.columns:
         past = data[variable]
